Backwards, progress? I think?

+++ Assets/Scripts/Network/Client/NetworkClientID.cs
@@ -3,10 +3,10 @@ using System.Collections;
 
 public class NetworkClientID
 {
-    public int ConnectionID;
+    public ulong ConnectionID;
     public uint NetID;
 
-    public NetworkClientID(int conID, uint netID)
+    public NetworkClientID(ulong conID, uint netID)
     {
         ConnectionID = conID;
         NetID = netID;
